Limpieza, reordenamiento del codigo y traducciones Añadidos algunos typecasts necesarios para evitar advertencias de conversion implicita. Agregada la implementacion de cierre del puerto serie Algunos parentesis se sacaron de preprocesamiento en las funciones Se agrego globaldefs en WiMOD HCI Layer.c y las modificaciones necesarias, adicionalmente algo necesario para la union con la version windows en globaldefs se agrego el comentario para Sistemas Unix y el comport por defecto en windows. Se tradujeron partes de algunos archivos a ingles

+++ source/CRC16.c
@@ -88,7 +88,7 @@ CRC16_Calc  (UINT8*             data,
     }
 
     // return result
-    return crc;
+    return (UINT16)crc;
 }
 #else
 {
